Image Contour Appplication - 2
1. CorrectConcave(imgfile)
If the object is not convex hull, by using cv2.convexHull(), figure out the convex hull contour of the object.

2. CircumscribeRect(imgfile):
Drawing having smallest area contour of circumscribing with the object.
Drawing errected contour of circumscribing with the object.

3. CircumscribingCircle(imgfile):
Drawing having smallest area circular contour  of circumscribing with the object.

4. CircumscribingEllipse(imgfile):
Drawing having smallest area ellipse contour  of circumscribing with the object.

5, Fitline(imgfile)
Drawing a line with minimized error. The error means distance from pixels and line.