Image Regularzation
If a histogram of certain image is concentrated at some values, the image will be seen as blurred.
So we have to regularize the image histogram, for example face detecting.

1.ImgRegul(imgfile)
 This function is for regularizing image histogram. This code is realize the fundamental of regularization.
Because this function does not use open cv function, it can deal with color image also(not only grayscale).

2. ImgRegul2(imgfile):
This function is using open cv function. So the codes are shorter, clearer. But the cv2.equalizeHist function does not support color image,
We can only put gray scale image to the parameter.