CLAHE(Contrast Limited Adaptive Histogram Equalization) method
cv2.equalizeHist() function only apply histogram regularization as a whole.
So if we wanna regularize dark image with an bright object, the contour of object is not easy to identify.
In this situation, we can use CLAHE method instead of cv2.equalizeHist.
CLAHE method use 8 by 8 kernels basically, we can modify the kernel size.
CLAHE does not regularize as a whle, instead a per kernel basis.
If image as some noise, it can be bigger when apply per kernel basis regularization.
But using CLAHE, we can ignore the noise expanding problem because contrast limiting reduces the noise.